Move NoDatabasesEmptyState into common
