+++ frontend/lint/module-boundaries.mjs
@@ -79,6 +79,7 @@ const elements = [
   createElement({ type: "shared", name: "i18n", enforceOutgoing: true }),
   createElement({ type: "shared", name: "metadata", enforceOutgoing: true }),
   createElement({ type: "shared", name: "metrics", enforceOutgoing: true }),
+  createElement({ type: "shared", name: "models", enforceOutgoing: true }),
   createElement({
     type: "shared",
     name: "metrics-viewer",
