+++ frontend/src/metabase/common/components/NoDatabasesEmptyState/index.tsx
@@ -1,0 +1,1 @@
+export { NoDatabasesEmptyState } from "./NoDatabasesEmptyState";
