+++ frontend/src/metabase/models/containers/NewModelOptions/NewModelOptions.tsx
@@ -2,10 +2,10 @@ import cx from "classnames";
 import type { Location } from "history";
 import { t } from "ttag";
 
+import { NoDatabasesEmptyState } from "metabase/common/components/NoDatabasesEmptyState";
 import CS from "metabase/css/core/index.css";
 import NewModelOption from "metabase/models/components/NewModelOption";
 import { useSelector } from "metabase/redux";
-import { NoDatabasesEmptyState } from "metabase/reference/databases/NoDatabasesEmptyState";
 import { getLearnUrl, getSetting } from "metabase/selectors/settings";
 import {
   canUserCreateNativeQueries,
